requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==17.2.0
autobahn==24.4.2
autodoc==0.3
Automat==0.6.0
beautifulsoup4==4.5.1
bleach==1.5.0
build==1.2.2.post1
certifi==2024.12.14
cffi==1.17.1
chardet==2.3.0
charset-normalizer==3.4.1
click==8.1.8
cobble==0.1.1
constantly==15.1.0
coverage==7.8.0
cryptography==1.7.1
cssselect==0.9.2
daphne==1.3.0
datefinder==0.7.3
DateTime==5.5
dateutils==0.6.6
decorator==5.1.1
Django==5.2.1
django-bootstrap-form==3.2.1
django-contrib-comments==1.7.3
django-crontab==0.7.1
django-extensions==1.7.4
django-localflavor==1.3
django-markdown-deux==1.0.5
django-mptt==0.8.7
django-nested-formset==0.1.4
django-rosetta==0.7.12
django-tagging==0.4.5
django-upgrade==1.22.2
django-xmlrpc==0.1.7
djangorestframework==3.16
djangorestframework_simplejwt==5.4.0
docutils==0.21.1
email_reply_parser==0.5.9
enum34==1.1.6
filebrowser-safe==0.4.6
fissix==24.4.24
flake8==7.1.1
funcsigs==1.0.2
future==0.16.0
grappelli-safe==0.4.5
hashids==1.1.0
html5lib==1.1
httplib2==0.9.2
hyperlink==21.0.0
idna==3.10
incremental==17.5.0
ipgetter2==1.1.10
lxml==5.3.0
mammoth==1.9.0
markdown2==2.3.2
mccabe==0.7.0
microsofttranslator==0.8
mock==2.0.0
modernize==0.8.0
mots-vides==2015.5.11
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==3.2.2
packaging==16.8
parsimonious==0.10.0
paypalrestsdk==1.12.0
pbr==6.1.0
pdfminer.six==20240706
pdfquery==0.4.3
pillow==11.2.1
pip-tools==7.4.1
polib==1.0.7
pre-commit==4.2.0
pyasn1==0.1.9
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodomex==3.4.2
pydot==1.2.3
pyflakes==3.2.0
PyJWT==1.4.2
PyMySQL==1.1.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
pypdf==5.5.0
pyproject_hooks==1.2.0
pyquery==2.0.1
pyth==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-linkedin==4.1
python-openid==2.2.5
pytz==2024.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==0.6.2
roman==2.0.0
rsa==3.4.2
setuptools==75.8.0
simplejson==3.10.0
six==1.10.0
soupsieve==2.6
sqlparse==0.5.3
tokenize_rt==6.1.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==1.3
tzdata==2025.2
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==2.3.0
validate_email==1.3
waitress==0.9.0
weasyprint==65.1
webencodings==0.5.1
WebOb==1.6.1
WebTest==3.0.2
Werkzeug==0.11.15
wheel==0.45.1
zope.interface==7.2
